Create initial files for RGB lamp
This is the first check-in for an Arduino-driven RGB lamp. The LEDs are on a strip
and are lit based on the amount of current flowing through the individual LEDs: more current = more brightness.

This initial check-in runs a simple program (cycle through the rainbow), using pre-calculated tweening steps
to create a more gradual transition.