Make tweening smarter
Before, the tweening logic was calculated in advance and added to the program queue.
This led to a number of issues with the size of the program and the loss of resolution
as more frames were added. (e.g., moving 20 units across 30 frames meant you would
go 0,0,0, ... 20)

Now the current tween state is calculated whenever there's not an explicit program
step to do. The logic figures out the timestamp for where it's been, the timestamp
for where it's going to, and uses the distance from start as a percentage that is then
applied to how far each of the colors needs to move towards their end goal.